+++ src/modules/remoteExplorer/treeDataProvider.ts
@@ -69,11 +69,13 @@ export default class RemoteTreeData
   private _rootsMap: Map<Id, ExplorerRoot> | null;
   private _map: Map<vscode.Uri['query'], ExplorerItem>;
 
-  private _onDidChangeFolder: vscode.EventEmitter<ExplorerItem> = new vscode.EventEmitter<
-    ExplorerItem
-  >();
+  private _onDidChangeFolder: vscode.EventEmitter<
+    ExplorerItem | undefined | null | void
+  > = new vscode.EventEmitter<ExplorerItem | undefined | null | void>();
   private _onDidChangeFile: vscode.EventEmitter<vscode.Uri> = new vscode.EventEmitter<vscode.Uri>();
-  readonly onDidChangeTreeData: vscode.Event<ExplorerItem> = this._onDidChangeFolder.event;
+  readonly onDidChangeTreeData: vscode.Event<
+    ExplorerItem | undefined | null | void
+  > = this._onDidChangeFolder.event;
   readonly onDidChange: vscode.Event<vscode.Uri> = this._onDidChangeFile.event;
 
   async refresh(item?: ExplorerItem): Promise<any> {